+++ rust/operator-binary/src/controller.rs
@@ -3,6 +3,7 @@ use crate::{
     authentication::{TrinoAuthenticationConfig, TrinoAuthenticationTypes},
     catalog::{config::CatalogConfig, FromTrinoCatalogError},
     command,
+    operations::{add_graceful_shutdown_config, graceful_shutdown_config_properties},
     product_logging::{get_log_properties, get_vector_toml, resolve_vector_aggregator_address},
 };
 
@@ -613,6 +614,8 @@ fn build_rolegroup_config_map(
                 dynamic_resolved_config
                     .insert(DISCOVERY_URI.to_string(), Some(discovery.discovery_uri()));
 
+                dynamic_resolved_config.extend(graceful_shutdown_config_properties(trino, role));
+
                 // The log format used by Trino
                 dynamic_resolved_config.insert(LOG_FORMAT.to_string(), Some("json".to_string()));
                 // The path to the log file used by Trino
@@ -701,21 +704,39 @@ fn build_rolegroup_config_map(
         }
     }
 
-    if let Some(opa_connect) = opa_connect_string {
-        let mut opa_config = BTreeMap::new();
-        opa_config.insert(
+    let access_control_properties = match opa_connect_string {
+        Some(opa_connect_string) => BTreeMap::from([
+            (
+                "access-control.name".to_string(),
+                Some("tech.stackable.trino.opa.OpaAuthorizer".to_string()),
+            ),
+            (
+                "opa.policy.uri".to_string(),
+                Some(opa_connect_string.to_string()),
+            ),
+        ]),
+        // The default access control with the name "default" allows everything but
+        // impersonation and graceful shutdown. As we need the user "admin" to have the permissions
+        // to gracefully shut down workers we default to the use of "allow-all".
+        //
+        // Not using OPA is meant as a testing environment anyway, as there is no authorization at all,
+        // so this elevation of privileges is acceptable compared for the need for graceful shutdown.
+        //
+        // If this however imposes a problem we could
+        // a.) let the user use OPA
+        // b.) or write a custom Trino Authorizer that behaves the same way the "default" one does,
+        //     with the difference, that the user "admin" is allowed to trigger a graceful shutdown.
+        None => BTreeMap::from([(
             "access-control.name".to_string(),
-            Some("tech.stackable.trino.opa.OpaAuthorizer".to_string()),
-        );
-        opa_config.insert("opa.policy.uri".to_string(), Some(opa_connect.to_string()));
+            Some("allow-all".to_string()),
+        )]),
+    };
 
-        let config_properties =
-            product_config::writer::to_java_properties_string(opa_config.iter())
-                .context(FailedToWriteJavaPropertiesSnafu)?;
+    let config_properties =
+        product_config::writer::to_java_properties_string(access_control_properties.iter())
+            .context(FailedToWriteJavaPropertiesSnafu)?;
 
-        cm_conf_data.insert(ACCESS_CONTROL_PROPERTIES.to_string(), config_properties);
-    }
-
+    cm_conf_data.insert(ACCESS_CONTROL_PROPERTIES.to_string(), config_properties);
     cm_conf_data.insert(JVM_CONFIG.to_string(), jvm_config.to_string());
 
     let jvm_sec_props: BTreeMap<String, Option<String>> = config
@@ -864,6 +885,7 @@ fn build_rolegroup_statefulset(
         &mut cb_prepare,
         &mut cb_trino,
     );
+    add_graceful_shutdown_config(trino, trino_role, &mut pod_builder, &mut cb_trino);
 
     // Add the needed stuff for catalogs
     env.extend(
@@ -902,7 +924,13 @@ fn build_rolegroup_statefulset(
 
     let container_prepare = cb_prepare
         .image_from_product_image(resolved_product_image)
-        .command(vec!["/bin/bash".to_string(), "-c".to_string()])
+        .command(vec![
+            "/bin/bash".to_string(),
+            "-x".to_string(),
+            "-euo".to_string(),
+            "pipefail".to_string(),
+            "-c".to_string(),
+        ])
         .args(vec![prepare_args.join(" && ")])
         .add_volume_mount("data", DATA_DIR_NAME)
         .add_volume_mount("rwconfig", RW_CONFIG_DIR_NAME)
@@ -920,7 +948,13 @@ fn build_rolegroup_statefulset(
 
     let container_trino = cb_trino
         .image_from_product_image(resolved_product_image)
-        .command(vec!["/bin/bash".to_string(), "-c".to_string()])
+        .command(vec![
+            "/bin/bash".to_string(),
+            "-x".to_string(),
+            "-euo".to_string(),
+            "pipefail".to_string(),
+            "-c".to_string(),
+        ])
         .args(command::container_trino_args(
             trino_authentication_config,
             catalogs,