+++ rust/operator-binary/src/operations/mod.rs
@@ -1,0 +1,3 @@
+mod graceful_shutdown;
+
+pub use graceful_shutdown::*;